Rotas e controllers

+++ src/Controllers/DishController.js
@@ -2,37 +2,47 @@ const knex = require("../database/knex");
 
 const sqliteConnection = require("../database/sqlite");
 
+const AppError = require("../utils/AppError");
+
+
 class DishController {
     async create(request, response) {
         const { name, description, ingredients, category, price } = request.body;
-        const { adm_id } = request.params;
+        const user_id = request.user.id;
+
+        const user = await knex("users").where("id", user_id).first();
+
+        if (!user || !user.is_admin) {
+            throw new AppError("Você não tem permissão para isso.", 401);
+        }
 
         const [dish_id] = await knex("dish").insert({
             name,
             description,
             category,
             price,
-            adm_id
+            adm_id: user_id
         })
 
         const ingredientsInsert = ingredients.map(name => {
             return {
-                adm_id,
+                adm_id: user_id,
                 dish_id,
                 name
             }
 
         })
+        
         await knex("ingredients").insert(ingredientsInsert);
 
-        response.json();
+        response.json({dish_id});
     }
 
     async show(request, response) {
         const { id } = request.params;
 
         const dish = await knex("dish").where({ id }).first();
-        const ingredients = await knex("ingredients").where({ dish_id: id }).orderBy("name");
+        const ingredients = await knex("ingredients").where({ dish_id : id }).orderBy("name");
 
         return response.json({
             ...dish,
@@ -43,14 +53,21 @@ class DishController {
     async delete(request, response) {
         const { id } = request.params;
 
+        const user = await knex("users").where("id", user_id).first();
+
+        if (!user || !user.is_admin) {
+            throw new AppError("Você não tem permissão para isso.", 401);
+        }
+
         await knex("dish").where({ id }).delete();
 
         return response.json()
     }
 
     async index(request, response) {
-        const { adm_id, name, ingredients } = request.query;
-
+        const { name, ingredients } = request.query;
+        const user_id = request.user.id;
+        const adm_id = user_id;
         let dish;
 
         if (ingredients) {
@@ -74,12 +91,12 @@ class DishController {
                 .orderBy("name");
         }
 
-        const userAdmIngredients = await knex("ingredients").where({ adm_id });
+        const userIngredients = await knex("ingredients").where({ adm_id });
         const dishWithIngredients = dish.map(dish => {
-            const noteTags = userAdmIngredients.filter(ingredient => ingredient.dish_id === dish.id)
+            const dishIngredients = userIngredients.filter(ingredient => ingredient.dish_id === dish.id)
             return {
-                ...noteTags,
-                tags: noteTags
+                ...dishIngredients,
+                ingredients: dishIngredients
             }
 
         })
@@ -88,18 +105,26 @@ class DishController {
     }
 
     async update(request, response) {
-        const { id } = request.params;
+        const user_id = request.user.id;
+
+        const adm_id = user_id;
 
         const { name, description, category, price } = request.body;
 
         const database = await sqliteConnection();
 
-        const dish = await database.get("SELECT * FROM dish WHERE id = (?)", [id]);
+        const dish = await database.get("SELECT * FROM dish WHERE id = (?)", [adm_id]);
 
-        dish.name = name;
-        dish.description = description;
-        dish.category = category;
-        dish.price = price;
+        const user = await knex("users").where("id", adm_id).first();
+
+        if (!user || !user.is_admin) {
+            throw new AppError("Você não tem permissão para isso.", 401);
+        }
+
+        dish.name = name ?? dish.name;
+        dish.description = description ?? dish.description;
+        dish.category = category ?? dish.category;
+        dish.price = price ?? dish.price;
 
         await database.run(`
             UPDATE dish SET
@@ -109,10 +134,10 @@ class DishController {
             price = ?,
             updated_at = DATETIME('now')
             WHERE id = ?`,
-            [dish.name, dish.description, dish.category, dish.price,  id]
-          );
-      
-          return response.json();
+            [dish.name, dish.description, dish.category, dish.price, adm_id]
+        );
+
+        return response.json();
 
     }
